// src/app/main/authentication/models/role.ts
export enum Role {

    //
    USER = 'USER',
    // ADMINISTRATION
    PDG = 'PDG',
    DGA = 'DGA',
    DAF = 'DAF',
    MR = 'MR',
    AUDIT_INTERNE= 'AUDIT_INTERNE',
    CDG= 'CDG',
    JURIDIQUE= 'JURIDIQUE',


    // ACHAT
    RESP_ACHAT = 'RESP_ACHAT',
    ASSIST_ACHAT = 'ASSIST_ACHAT',

    // RH
    RESP_RH = 'RESP_RH',
    ASSIST_RH = 'ASSIST_RH',


    // LOGISTIQUE
    RESP_LOG = 'RESP_LOG',
    ASSIST_LOG = 'ASSIST_LOG',


    // SI
    ADMIN = 'ADMIN',
    RESP_SI = 'RESP_SI',
    ASSIST_SI = 'ASSIST_SI',


    // MARCHE
    RESP_MARCHE = 'RESP_MARCHE',
    ASSIST_MARCHE = 'ASSIST_MARCHE',

    // COMPTABILITE
    RESP_COMPTA = 'RESP_COMPTA',
    ASSIST_COMPTA = 'ASSIST_COMPTA',


    // CAISSE
    RESP_CAISSE= 'RESP_CAISSE',
    CAISSIER= 'CAISSIER',

    // TECHNIQUE
    DIR_TECH = 'DIR_TECH',
    RESP_METREE = 'RESP_METREE',
    METREUR = 'METREUR',
    RESP_ST = 'RESP_ST',
    // ASSIST_ST = 'ASSIST_ST',
    TOPOGRAPHE = 'TOPOGRAPHE',
    TRACEUR = 'TRACEUR',

    // CHANTIER

    DIR_PROJET = 'DIR_PROJET',
    CHEF_ZONE = 'CHEF_ZONE',
    ING_CH = 'ING_CH',
    COND_CH = 'COND_CH',
    TECH_CH = 'TECH_CH',
    ASSIST_TECH_CH = 'ASSIST_TECH_CH',
    CHEF_CH = 'CHEF_CH',
    COMPT_CH = 'COMPT_CH',
    POINTEUR = 'POINTEUR',
    CAPORAL = 'CAPORAL',
    MAGASINIER_CH = 'MAGASINIER_CH',
    GARDIENT_CH_JOUR = 'GARDIENT_CH_JOUR',
    GARDIENT_CH_NUIT = 'GARDIENT_CH_NUIT',

    // DEPOT
    RESP_DEPOT = 'RESP_DEPOT',

}